+++ src/vanguards/main.py
@@ -166,6 +166,10 @@ def control_loop(state):
                    functools.partial(vanguards.VanguardState.new_consensus_event,
                                      state, controller),
                                     stem.control.EventType.NEWCONSENSUS)
+    controller.add_event_listener(
+                   functools.partial(vanguards.VanguardState.signal_event,
+                                     state, controller),
+                                    stem.control.EventType.SIGNAL)
 
   # Blah...
   while controller.is_alive():